add summarizer and smart quiz frontend part
